requirements.txt
Django==3.1.12
django-cors-headers==3.8.0
djangorestframework==3.12.4
django-heroku
Pillow==8.2.0
gunicorn==20.1.0
psycopg2